requirements.txt
# TTI Sensor Analysis System - Requirements
numpy>=1.24.0
Pillow>=10.0.0
Flask>=3.0.0
Werkzeug>=3.0.0
gunicorn>=21.2.0
